+++ dsptoolbox/_standard.py
@@ -32,9 +32,9 @@ def _welch(x, y, fs_hz: int, window_type: str = 'hann',
     Parameters
     ----------
     x : `np.ndarray`
-        First signal
+        First signal with shape (time samples, channel).
     y : `np.ndarray`
-        Second signal
+        Second signal with shape (time samples, channel).
     fs_hz : int
         Sampling rate in Hz.
     window_type : str, optional
@@ -62,6 +62,8 @@ def _welch(x, y, fs_hz: int, window_type: str = 'hann',
         Complex cross spectral density vector if x and y are different.
         Alternatively, the (real) autocorrelation power spectral density when
         x and y are the same. If density or spectrum depends on scaling.
+        Depending on the input, the output shape is (time samples) or
+        (time samples, channel).
 
     References
     ----------
@@ -74,19 +76,28 @@ def _welch(x, y, fs_hz: int, window_type: str = 'hann',
       See http://arxiv.org/abs/gr-qc/0509116.
 
     """
-    # from time import time
     if type(x) != np.ndarray:
-        x = np.array(x).squeeze()
+        x = np.asarray(x).squeeze()
     if type(y) != np.ndarray:
-        y = np.array(y).squeeze()
+        y = np.asarray(y).squeeze()
     assert x.shape == y.shape, \
         'Shapes of data do not match'
-    assert len(x.shape) < 2, f'{x.shape} are too many dimensions. Use flat' +\
-        ' arrays instead'
-    valid_window_sizes = np.array([int(2**x) for x in range(4, 17)])
+    # NOTE: Computing the spectrum in a vectorized manner for all channels
+    # simultaneously does not seem to be faster than doing it sequentally
+    # for each channel. Maybe parallelizing with something like numba could
+    # be advantageous...
+    if x.ndim == 2:
+        multi_channel = True
+    else:
+        multi_channel = False
+
+    assert len(x.shape) <= 2, f'{x.shape} are too many dimensions. Use flat' +\
+        ' arrays or 2D-Arrays instead'
+
+    valid_window_sizes = np.array([int(2**x) for x in range(3, 19)])
     assert window_length_samples in valid_window_sizes, \
-        'Window length should be a power of 2 between [16, 65536] or ' +\
-        '[2**4, 2**16]'
+        'Window length should be a power of 2 between [8, 262_144] or ' +\
+        '[2**3, 2**18]'
     assert overlap_percent >= 0 and overlap_percent < 100, \
         'overlap_percent should be between 0 and 100'
     valid_average = ['mean', 'median']
@@ -97,6 +108,8 @@ def _welch(x, y, fs_hz: int, window_type: str = 'hann',
     assert scaling in valid_scaling, f'{scaling} is not valid. Use ' +\
         'power spectrum, power spectral density, amplitude spectrum, ' +\
         'amplitude spectral density or None'
+    if scaling is None:
+        scaling = ''
 
     # Window and step
     window = windows.get_window(
@@ -109,24 +122,20 @@ def _welch(x, y, fs_hz: int, window_type: str = 'hann',
         warn('Selected window type and overlap do not meet the constant ' +
              'overlap and add constraint! Results might be distorted')
 
-    # Start Parameters
-    n_frames, padding_samp = \
-        _compute_number_frames(window_length_samples, step, len(x))
-    x = _pad_trim(x, len(x) + padding_samp)
-    y = _pad_trim(y, len(y) + padding_samp)
-    x_frames = np.zeros((window_length_samples, n_frames), dtype='float')
-    y_frames = np.zeros_like(x_frames)
+    if not multi_channel:
+        x = x[..., None]
+        y = y[..., None]
 
-    # Create time frames
-    start = 0
-    for n in range(n_frames):
-        x_frames[:, n] = x[start:start+window_length_samples].copy()
-        y_frames[:, n] = y[start:start+window_length_samples].copy()
-        start += step
+    x_frames = _get_framed_signal(x, window_length_samples, step)
+    y_frames = _get_framed_signal(y, window_length_samples, step)
 
     # Window
-    x_frames *= window[..., None]
-    y_frames *= window[..., None]
+    x_frames *= window[:, np.newaxis, np.newaxis]
+    y_frames *= window[:, np.newaxis, np.newaxis]
+
+    if not multi_channel:
+        x_frames = np.squeeze(x_frames)
+        y_frames = np.squeeze(y_frames)
 
     # Detrend
     if detrend:
@@ -139,10 +148,10 @@ def _welch(x, y, fs_hz: int, window_type: str = 'hann',
 
     # Direct averaging much faster
     if average == 'mean':
-        csd = np.mean(sp_frames, axis=-1)
+        csd = np.mean(sp_frames, axis=1)
     else:
-        csd = np.median(sp_frames.real, axis=-1) + 1j * \
-            np.median(sp_frames.imag, axis=-1)
+        csd = np.median(sp_frames.real, axis=1) + 1j * \
+            np.median(sp_frames.imag, axis=1)
         # Bias according to reference
         n = sp_frames.shape[1] if sp_frames.shape[1] % 2 == 1 else \
             sp_frames.shape[1] - 1
@@ -161,11 +170,11 @@ def _welch(x, y, fs_hz: int, window_type: str = 'hann',
 
     # Zero frequency fix when detrending (especially useful for dB plotting)
     if detrend:
-        csd[0] = csd[1]
+        csd[0, ...] = csd[1, ...]
 
     csd *= factor
-    csd[0] /= 2
-    csd[-1] /= 2
+    csd[0, ...] /= 2
+    csd[-1, ...] /= 2
 
     if 'amplitude' in scaling:
         csd = np.sqrt(csd)
@@ -173,7 +182,6 @@ def _welch(x, y, fs_hz: int, window_type: str = 'hann',
     # Cast to real output if there is no imaginary part
     if np.all(csd.imag == 0):
         csd = csd.real
-
     return csd
 
 
@@ -235,7 +243,7 @@ def _minimum_phase(magnitude: np.ndarray, unwrapped: bool = True):
 def _stft(x: np.ndarray, fs_hz: int, window_length_samples: int = 2048,
           window_type: str = 'hann', overlap_percent=50,
           fft_length_samples: int = None, detrend: bool = True,
-          padding: bool = True, scaling: bool = False):
+          padding: bool = False, scaling: bool = False):
     """Computes the STFT of a signal. Output matrix has (freqs_hz, seconds_s).
 
     Parameters
@@ -260,7 +268,8 @@ def _stft(x: np.ndarray, fs_hz: int, window_length_samples: int = 2048,
         Detrending from each time segment (removing mean). Default: True.
     padding : bool, optional
         When `True`, the original signal is padded in the beginning and ending
-        so that no energy is lost due to windowing. Default: `True`.
+        so that no energy is lost due to windowing when the COLA constraint is
+        met. Default: `False`.
     scaling : bool, optional
         When `True`, the output is scaled as an amplitude spectrum, otherwise
         no scaling is applied. See references for details. Default: `False`.
@@ -272,7 +281,7 @@ def _stft(x: np.ndarray, fs_hz: int, window_length_samples: int = 2048,
     freqs_hz : `np.ndarray`
         Frequency vector.
     stft : `np.ndarray`
-        STFT matrix with shape (frequency, time).
+        STFT matrix with shape (frequency, time, channel).
 
     References
     ----------
@@ -301,23 +310,11 @@ def _stft(x: np.ndarray, fs_hz: int, window_length_samples: int = 2048,
 
     # Padding
     if padding:
-        x = _pad_trim(x, len(x)+overlap_samples, in_the_end=False)
-        x = _pad_trim(x, len(x)+overlap_samples, in_the_end=True)
-
-    # Number of samples and padding
-    n_frames, padding_samp = \
-        _compute_number_frames(window_length_samples, step, len(x))
-
-    x = _pad_trim(x, len(x) + padding_samp)
-    time_x = np.zeros((window_length_samples, n_frames), dtype='float')
-
-    # Create time frames
-    start = 0
-    for n in range(n_frames):
-        time_x[:, n] = x[start:start+window_length_samples].copy()
-        start += step
+        x = np.pad(x, ((overlap_samples, overlap_samples), (0, 0)))
+    # Framed signal
+    time_x = _get_framed_signal(x, window_length_samples, step, True)
     # Windowing
-    time_x *= window[..., None]
+    time_x *= window[..., np.newaxis, np.newaxis]
     # Detrend
     if detrend:
         time_x -= np.mean(time_x, axis=0)
@@ -381,6 +378,14 @@ def _csm(time_data: np.ndarray, sampling_rate_hz: int,
       comprehensive list of window functions and some new at-top windows.
 
     """
+    # ===== Remarks on speed =============
+    # It has been tried to vectorize the whole computation of the CSM by using
+    # a multi-channel approach in the _welch() function or with a class. This
+    # leads to a dramatic drop in performance and an elevated memory cost.
+    # Maybe using some parallel computing framework like numba would make it
+    # faster, but less readable and it would be a new dependency of the
+    # package... So far the double loop has been best solution
+    # =====================================
     number_of_channels = time_data.shape[1]
     csm = np.zeros((window_length_samples//2+1,
                     number_of_channels,
@@ -389,7 +394,6 @@ def _csm(time_data: np.ndarray, sampling_rate_hz: int,
         for ind2 in range(ind1, number_of_channels):
             # Complex conjugate second signal and not first (like transposing
             # the matrix)
-            # csm[:, ind1, ind2] = \
             csm[:, ind2, ind1] = \
                 _welch(time_data[:, ind1],
                        time_data[:, ind2],
@@ -405,180 +409,6 @@ def _csm(time_data: np.ndarray, sampling_rate_hz: int,
     csm += np.swapaxes(csm, 1, 2).conjugate()
     f = np.fft.rfftfreq(window_length_samples, 1/sampling_rate_hz)
     return f, csm
-
-
-# =============================================================================
-# It was shown that this helper class was not faster than the _csm function.
-# Vectorizing some of the computations did not seem to bring any advantages
-# to the speed. Especially, np.fft.rfft was a little slower and np.unwrap was
-# much slower when applied to large matrices along specific dimensions...
-# =============================================================================
-# class _CSMHelper():
-#     """This is a helper class to compute the csm in an efficient manner.
-
-#     """
-#     def __init__(self, time_data: np.ndarray, sampling_rate_hz: int,
-#                  window_length_samples: int = 1024,
-#                  window_type: str = 'hann', overlap_percent: int = 50,
-#                  detrend: bool = True, average: str = 'mean',
-#                  scaling: str = 'power'):
-#         """Computes the cross spectral matrix of a multichannel signal.
-#         Output matrix has (frequency, channels, channels).
-
-#         Parameters
-#         ----------
-#         time_data : `np.ndarray`
-#             Signal
-#         sampling_rate_hz : int
-#             Sampling rate in Hz.
-#         window_length_samples : int, optional
-#             Window length to be used. Determines frequency resolution in the
-#             end. Only powers of 2 are accepted. Default: 1024.
-#         window_type : str, optional
-#             Window type to be used. Refer to scipy.signal.windows for
-#             available ones. Default: `'hann'`
-#         overlap_percent : int, optional
-#             Overlap in percentage. Default: 50.
-#         detrend : bool, optional
-#             Detrending from each time segment (removing mean). Default: True.
-#         average : str, optional
-#             Type of mean to be computed. Take `'mean'` or `'np.median'`.
-#             Default: `'mean'`
-#         scaling : str, optional
-#             Scaling for spectral power density or spectrum. Takes `'power'`
-#             or `'spectrum'`. Default: `'power'`.
-
-#         Attributes
-#         ----------
-#         All passed parameters are saved as attributes.
-
-#         - `get_csm()`: Method to trigger the computation of csm.
-
-#         """
-#         assert time_data.shape[1] > 1, \
-#             'There has to be at least 2 channels to compute the csm'
-#         valid_window_sizes = np.array([int(2**x) for x in range(7, 17)])
-#         assert window_length_samples in valid_window_sizes, \
-#             'Window length should be a power of 2 between [128, 65536] ' +\
-#             'or [2**7, 2**16]'
-#         assert overlap_percent > 0 and overlap_percent < 100, \
-#             'overlap_percent should be between 0 and 100'
-#         valid_average = ['mean', 'median']
-#         assert average in valid_average, f'{average} is not valid. Use ' +\
-#             'either mean or median'
-#         valid_scaling = ['power', 'spectrum']
-#         assert scaling in valid_scaling, f'{scaling} is not valid. Use ' +\
-#             'either power or spectrum'
-
-#         # Window and step
-#         window = windows.get_window(
-#             window_type, window_length_samples, fftbins=True)
-#         overlap_samples = int(overlap_percent/100 * window_length_samples)
-#         step = window_length_samples - overlap_samples
-
-#         # Check COLA
-#         assert check_COLA(
-#             window, nperseg=len(window), noverlap=overlap_samples),\
-#             'Selected window type and overlap do not meet the constant ' +\
-#             'overlap and add constraint. Please select other.'
-
-#         # Save all attributes in object
-#         self.time_data = time_data
-#         self.window = window
-#         self.window_length_samples = window_length_samples
-#         self.sampling_rate_hz = sampling_rate_hz
-#         self.overlap_samples = overlap_samples
-#         self.step = step
-#         self.detrend = detrend
-
-#         if average == 'mean':
-#             self.average_func = np.mean
-#         else:
-#             self.average_func = np.median
-
-#         if scaling == 'power':
-#             self.factor = 2 / (window @ window) / sampling_rate_hz
-#         else:
-#             self.factor = 2 / sum(window)**2 / sampling_rate_hz
-#         self.f_vec = \
-#             np.fft.rfftfreq(window_length_samples, 1/sampling_rate_hz)
-
-#     def _compute_framed_spectra(self):
-#         """Computes the framed spectra.
-
-#         """
-#         # Compute number of needed frames and padding
-#         original_length = self.time_data.shape[0]
-#         n_frames, padding_samp = _compute_number_frames(
-#             self.window_length_samples, self.step, original_length)
-#         self.time_data = _pad_trim(
-#             self.time_data, original_length + padding_samp)
-#         td_frames = np.zeros(
-#             (self.window_length_samples, n_frames, self.time_data.shape[1]),
-#             dtype='float')
-#         self.n_frames = n_frames
-
-#         # Create time frames (time samples, frame, channel)
-#         start = 0
-#         for n in range(n_frames):
-#             td_frames[:, n, :] = \
-#                 self.time_data[
-#                     start:start+self.window_length_samples, :].copy()
-#             start += self.step
-
-#         # Window
-#         td_frames *= self.window[:, np.newaxis, np.newaxis]
-#         # Detrend
-#         if self.detrend:
-#             td_frames -= np.mean(td_frames, axis=0)
-#         # Spectra (frequency bins, frames, channel)
-#         self.sp_frames = np.fft.rfft(td_frames, axis=0)
-
-#     def get_csm(self):
-#         """Computes and returns the CSM.
-
-#         Returns
-#         -------
-#         f : `np.ndarray`
-#             Frequency vector.
-#         csm : `np.ndarray`
-#             Cross-spectral density matrix with shape
-#             (frequency, channel, channel).
-
-#         """
-#         self._compute_framed_spectra()
-#         number_of_channels = self.time_data.shape[1]
-#         csm_framed = np.zeros((self.window_length_samples//2+1,
-#                                self.n_frames,
-#                                number_of_channels,
-#                                number_of_channels), dtype=np.complex64)
-
-#         # Maybe second loop can be vectorized by using np.roll along one
-#         # channel axis. There would be unnecessary computations but it
-#         # might be faster
-#         for ind1 in range(number_of_channels):
-#             for ind2 in range(ind1, number_of_channels):
-#                 csm_framed[:, :, ind1, ind2] = \
-#                     np.conjugate(self.sp_frames[:, :, ind1]) * \
-#                     self.sp_frames[:, :, ind2]
-#                 if ind1 == ind2:
-#                     csm_framed[:, :, ind1, ind2] *= 0.5
-
-#         # Get magnitude and phase
-#         magnitude = np.abs(csm_framed)
-#         # phase = np.angle(csm_framed)
-#         phase = np.unwrap(np.angle(csm_framed), axis=0)
-#         del csm_framed
-#         # Average along frames (axis=1)
-#         magnitude = self.average_func(magnitude, axis=1)
-#         phase = self.average_func(phase, axis=1)
-
-#         # Redo matrix with factor
-#         csm = magnitude * np.exp(1j*phase) * self.factor
-#         # csm = np.mean(csm_framed, axis=1) * self.factor
-#         # Complete lower triangle csm matrix
-#         csm += np.swapaxes(csm, 1, 2).conjugate()
-#         return self.f_vec, csm
 
 
 def _center_frequencies_fractional_octaves_iec(nominal, num_fractions):
@@ -712,28 +542,32 @@ def _kaiser_window_beta(A):
     return beta
 
 
-def _indexes_above_threshold_dbfs(time_vec: np.ndarray, threshold_dbfs: float,
-                                  attack_samples: int, release_samples: int):
-    """Returns indexes with power above a passed threshold (in dBFS) in a time
-    series. time_vec is normalized prior to computation.
+def _indices_above_threshold_dbfs(time_vec: np.ndarray, threshold_dbfs: float,
+                                  attack_smoothing_coeff: int,
+                                  release_smoothing_coeff: int,
+                                  normalize: bool = True):
+    """Returns indices with power above a given power threshold (in dBFS) in a
+    time series. time_vec can be normalized to peak value prior to computation.
 
     Parameters
     ----------
     time_vec : `np.ndarray`
-        Time series for which to find indexes above power threshold.
+        Time series for which to find indices above power threshold. Can only
+        take one channel.
     threshold_dbfs : float
-        Threshold to be used.
-    attack_samples : int
-        Number of samples representing attack time signal has surpassed
-        power threshold.
-    release_samples : int
-        Number of samples representing release time after signal has decayed
-        below power threshold.
+        Threshold in dBFS to be regarded for activation.
+    attack_smoothing_coeff : int
+        Coefficient for attack smoothing for level computation.
+    release_smoothing_coeff : int
+        Coefficient for release smoothing for level computation.
+    normalize : bool, optional
+        When `True`, signal is normalized such that the threshold is relative
+        to peak level and not absolute. Default: `True`.
 
     Returns
     -------
-    indexes_above : `np.ndarray`
-        Array of type boolean with length of time_vec indicating indexes
+    indices_above : `np.ndarray`
+        Array of type boolean with length of time_vec indicating indices
         above threshold with `True` and below with `False`.
 
     """
@@ -741,26 +575,27 @@ def _indexes_above_threshold_dbfs(time_vec: np.ndarray, threshold_dbfs: float,
     assert time_vec.ndim == 1, \
         'Function is implemented for 1D-arrays only'
 
-    # Find peak value index
-    max_ind = np.argmax(np.abs(time_vec))
+    # Normalization
+    if normalize:
+        time_vec /= np.abs(time_vec).max()
 
     # Power in dB
-    time_power = 20*np.log10(np.abs(time_vec))
+    time_power = time_vec.squeeze()**2
 
-    # Normalization
-    time_power -= time_power[max_ind]
+    momentary_gain = np.zeros(len(time_power))
+    for i in np.arange(1, len(time_power)):
+        if momentary_gain[i] > time_power[i-1]:
+            coeff = attack_smoothing_coeff
+        elif momentary_gain[i] < time_power[i-1]:
+            coeff = release_smoothing_coeff
+        else:
+            coeff = 0
+        momentary_gain[i] = coeff*time_power[i] + (1-coeff)*momentary_gain[i-1]
+    momentary_gain = 10*np.log10(momentary_gain)
 
-    # All indexes above threshold
-    indexes_above_0 = time_power > threshold_dbfs
-    indexes_above = np.zeros_like(indexes_above_0).astype(bool)
-
-    # Apply release and attack
-    for ind in np.arange(len(indexes_above)):
-        # Attack after certain amount of samples surpass threshold
-        ind_attack = 0 if ind-attack_samples < 0 else ind-attack_samples
-        if np.all(indexes_above_0[ind_attack:ind]):
-            indexes_above[ind:ind+release_samples] = True
-    return indexes_above
+    # Get Indices above threshold
+    indices_above = momentary_gain > threshold_dbfs
+    return indices_above
 
 
 def _detrend(time_data: np.ndarray, polynomial_order: int) -> np.ndarray:
@@ -787,28 +622,33 @@ def _detrend(time_data: np.ndarray, polynomial_order: int) -> np.ndarray:
     return time_data
 
 
-def _rms(x: np.ndarray) -> float:
-    """Root mean squared value of a discrete time series
+def _rms(x: np.ndarray) -> float | np.ndarray:
+    """Root mean squared value of a discrete time series.
 
     Parameters
     ----------
     x : `np.ndarray`
-        Time series
+        Time series.
 
     Returns
     -------
-    rms : float
-        Root mean squared value
+    rms : float or `np.ndarray`
+        Root mean squared of a signal. Float or np.ndarray depending on input.
 
     """
+    single_dim = False
     if x.ndim < 2:
+        single_dim = True
         x = x[..., None]
     elif x.ndim == 2:
         pass
     else:
         raise ValueError('Shape of array is not valid. Only 2D-Arrays ' +
                          'are valid')
-    return np.sqrt(np.mean(x**2, axis=0))
+    rms_vals = np.sqrt(np.mean(x**2, axis=0))
+    if single_dim:
+        rms_vals = np.squeeze(rms_vals)
+    return rms_vals
 
 
 def _get_framed_signal(td: np.ndarray, window_length_samples: int,
@@ -857,3 +697,88 @@ def _get_framed_signal(td: np.ndarray, window_length_samples: int,
     if not keep_last_frame:
         td_framed = td_framed[:, :-1, :]
     return td_framed
+
+
+def _reconstruct_framed_signal(td_framed: np.ndarray, step_size: int,
+                               window: str | np.ndarray = None,
+                               original_signal_length: int = None,
+                               safety_threshold: float = 1e-4) \
+        -> np.ndarray:
+    """Gets and returns a framed signal into its vector representation.
+
+    Parameters
+    ----------
+    td_framed : `np.ndarray`
+        Framed signal with shape (time samples, frame, channel).
+    step_size : int
+        Step size in samples between frames (also known as hop length).
+    window : str, `np.ndarray`, optional
+        Window (if applies). Pass `None` to avoid using a window during
+        reconstruction. Default: `None`.
+    original_signal_length : int, optional
+        When different than `None`, the output is padded or trimmed to this
+        length. Default: `None`.
+    safety_threshold : float, optional
+        When reconstructing the signal with a window, very small values can
+        lead to instabilities. This safety threshold avoids dividing with
+        samples beneath this value. Default: 1e-4.
+
+        Dividing by 1e-4 is the same as amplifying by 80 dB.
+
+    Returns
+    -------
+    td : `np.ndarray`
+        Reconstructed signal.
+
+    """
+    assert td_framed.ndim == 3, \
+        'Framed signal must contain exactly three dimensions'
+    if window is not None:
+        if type(window) == str:
+            window = windows.get_window(window, td_framed.shape[0])
+        elif type(window) == np.ndarray:
+            assert window.ndim == 1, \
+                'Window must be a 1D-array'
+            assert window.shape[0] == td_framed.shape[0], \
+                'Window length does not match signal length'
+        td_framed *= window[:, np.newaxis, np.newaxis]
+
+    total_length = int(step_size * td_framed.shape[1] +
+                       td_framed.shape[0]*(1 - step_size/td_framed.shape[0]))
+    td = np.zeros((total_length, td_framed.shape[-1]))
+
+    start = 0
+    for i in range(td_framed.shape[1]):
+        td[start:start+td_framed.shape[0], :] += td_framed[:, i, :]
+        start += step_size
+
+    if window is not None:
+        envelope = _get_window_envelope(
+            window, total_length, step_size,
+            td_framed.shape[1], True)
+        if safety_threshold is not None:
+            envelope = np.clip(envelope, a_min=safety_threshold, a_max=None)
+        non_zero = envelope > np.finfo(td.dtype).tiny
+        td[non_zero, ...] /= envelope[non_zero, np.newaxis]
+
+    if original_signal_length is not None:
+        td = _pad_trim(td, original_signal_length)
+    return td
+
+
+def _get_window_envelope(window: np.ndarray, total_length_samples: int,
+                         step_size_samples: int, number_frames: int,
+                         squared: bool = True):
+    """Compute the window envelope for a given window with step size and total
+    length. The window can be squared or not.
+
+    """
+    if squared:
+        window **= 2
+    envelope = np.zeros(total_length_samples)
+
+    start = 0
+    for _ in range(number_frames):
+        envelope[start:start+len(window)] += window
+        start += step_size_samples
+    return envelope